requirements.txt
fastapi==0.116.1
httpx==0.28.1
logics==1.10.4
pydantic==2.11.7
pytest==8.3.2
python-dotenv==1.1.1
uvicorn==0.35.0
